# swagger.yaml
openapi: "3.0.3"
info:
  title: MySlotMate API
  description: |
    Backend API for **MySlotMate** – a slot-booking platform with wallet-based payments,
    Razorpay payouts, Aadhar (Setu OKYC) verification, AWS S3 file uploads,
    and real-time inbox messaging.

    Every response follows the standard envelope:
    ```json
    { "success": true|false, "data": ..., "message": "...", "error": "..." }
    ```
  version: "2.0.0"
servers:
  - url: http://localhost:{port}
    description: Local development
    variables:
      port:
        default: "8080"
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSignUpRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/verify-aadhar/init:
    post:
      tags: [Auth]
      summary: Initiate Aadhar OKYC verification
      operationId: initiateAadhar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateAadharRequest"
      responses:
        "200":
          description: OTP sent — returns transaction ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_AadharInit"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/verify-aadhar/complete:
    post:
      tags: [Auth]
      summary: Complete Aadhar verification with OTP
      operationId: completeAadhar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteAadharRequest"
      responses:
        "200":
          description: Verification successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Users ───────────────────────────────────
  /users/me:
    get:
      tags: [Users]
      summary: Get own user profile
      operationId: getMyProfile
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [Users]
      summary: Update own profile
      operationId: updateProfile
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdateRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Saved Experiences ───────────────────────
  /users/saved-experiences:
    post:
      tags: [SavedExperiences]
      summary: Save/bookmark an experience
      operationId: saveExperience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveExperienceRequest"
      responses:
        "201":
          description: Experience saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_SavedExperience"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [SavedExperiences]
      summary: List all saved experiences
      operationId: getSavedExperiences
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Array of saved experiences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_SavedExperienceList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /users/saved-experiences/{eventID}:
    delete:
      tags: [SavedExperiences]
      summary: Remove a saved experience
      operationId: unsaveExperience
      parameters:
        - $ref: "#/components/parameters/EventIDPath"
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Experience unsaved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /users/saved-experiences/{eventID}/check:
    get:
      tags: [SavedExperiences]
      summary: Check if an event is saved
      operationId: isExperienceSaved
      parameters:
        - $ref: "#/components/parameters/EventIDPath"
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "Returns { saved: true/false }"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      saved:
                        type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Hosts ───────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: "Firebase ID token passed as `Authorization: Bearer <token>`. Admin endpoints additionally require the token's email to match the `ADMIN_EMAIL` env variable."

  # ── Parameters ──────────────────────────────
  parameters:
...
    EventIDPath:
      name: eventID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Event UUID
...
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ── Schemas ─────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string

    # ── Request Bodies ────────────────────────
    UserSignUpRequest:
      type: object
      required: [auth_uid, email, name, phn_number]
      properties:
        auth_uid:
          type: string
          description: Firebase Auth UID
          example: "firebase_uid_abc123"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "Ankit Sharma"
        phn_number:
          type: string
          example: "+919876543210"
    UserProfileUpdateRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
    InitiateAadharRequest:
      type: object
      required: [user_id, aadhar_number]
      properties:
        user_id:
          type: string
          format: uuid
        aadhar_number:
          type: string
          example: "123456789012"
    CompleteAadharRequest:
      type: object
      required: [user_id, transaction_id, otp]
      properties:
        user_id:
          type: string
          format: uuid
        transaction_id:
          type: string
        otp:
          type: string
          example: "123456"
    SaveExperienceRequest:
      type: object
      required: [user_id, event_id]
      properties:
        user_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        auth_uid:
          type: string
        name:
          type: string
        phn_number:
          type: string
        email:
          type: string
          format: email
        avatar_url:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        account_id:
          type: string
          format: uuid
          nullable: true
        is_verified:
          type: boolean
        verified_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SavedExperience:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        saved_at:
          type: string
          format: date-time
...
    SuccessResponse_User:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/User"
...
    SuccessResponse_SavedExperience:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/SavedExperience"
    SuccessResponse_SavedExperienceList:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/SavedExperience"
...
    SuccessResponse_Message:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
              example: "operation successful"
    SuccessResponse_AadharInit:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            transaction_id:
              type: string
            message:
              type: string
              example: "OTP sent to registered mobile"